requirements.txt
numpy==1.26.4
pandas==2.2.2
pyaes==1.6.1
pyasn1==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
rsa==4.9
six==1.16.0
Telethon==1.35.0
tzdata==2024.1
